telegram-upload-bot/requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1

# Optional for future HTTP calls
requests==2.32.3
